course-code/labs/lab-01/solution/rag/requirements.txt
fastembed
faiss-cpu
fastapi[standard]
prometheus-client
numpy
